requirements.txt
discord.py==1.3.3
Flask==1.1.2
requests==2.22.0
websockets==8.1